add expressions table

+++ server/database/migrations/2017_11_11_082103_create_table_instructions.php
@@ -27,7 +27,7 @@ class CreateTableInstructions extends Migration
             $table->increments('id');
             $table->integer('instruction_id')->index();
             $table->tinyInteger('pos')->index();
-            $table->tinyInteger('size');
+            $table->integer('size');
             $table->string('type');
 
             $table->string('reg')->nullable();
